Add dag run details page.
